+++ back-end/src/main/java/com/utd/se3345/project1/sortingapp/algorithms/InsertionSort.java
@@ -2,8 +2,8 @@ package com.utd.se3345.project1.sortingapp.algorithms;
 
 public class InsertionSort implements Sort {
 	
-	private int comparisons;
-	private int movements;
+	private long comparisons;
+	private long movements;
 
 	@Override
 	public void sort(int[] arr) {
@@ -24,11 +24,11 @@ public class InsertionSort implements Sort {
 		
 	}
 
-	public int getComparisons() {
+	public long getComparisons() {
 		return comparisons;
 	}
 
-	public int getMovements() {
+	public long getMovements() {
 		return movements;
 	}
