+++ back-end/src/main/java/com/utd/se3345/project1/sortingapp/algorithms/Sort.java
@@ -3,7 +3,7 @@ package com.utd.se3345.project1.sortingapp.algorithms;
 public interface Sort {
 	
 	public void sort(int[] arr);
-	public int getComparisons();
-	public int getMovements();
+	public long getComparisons();
+	public long getMovements();
 
 }
